+++ src/RutokenPkcs11Interop/RutokenPkcs11Interop/HighLevelAPI81/Factories/VolumeFormatInfoExtendedFactory.cs
@@ -12,8 +12,8 @@ namespace RutokenPkcs11Interop.HighLevelAPI81.Factories
     /// </summary>
     public class VolumeFormatInfoExtendedFactory : IVolumeFormatInfoExtendedFactory
     {
-        public IVolumeFormatInfoExtended Create(uint volumeSize, FlashAccessMode accessMode,
-            CKU volumeOwner, uint flags)
+        public IVolumeFormatInfoExtended Create(ulong volumeSize, FlashAccessMode accessMode,
+            CKU volumeOwner, ulong flags)
         {
             return new VolumeFormatInfoExtended(volumeSize, accessMode, volumeOwner, flags);
         }
